WIP: Renamed ValidatorError, added tracebacks.
The validatorError is now called a ValidatorLogEntry
(since it is not an exception).

These entries can now also accumulate a traceback list,
to make debugging easier.

Whether or not tracebacks will be accumulated will depend
on the IIIFValidator.debug boolean. Also renamed the booleans
for logging errors and warnings to collect_errors and
collect_warnings, in order to differentiate them from the
functions that actually log errors and warnings.

Fixed #3