New: Added _Path class to handle paths.
This class will be responsible for representing the
paths within documents. This job has become sufficiently
nuanced that a class is needed.

This class will need to be able to compare itself to another
while ignoring indexes (for hashing), represent itself, and
will likely have a function similar to 'endswith' to see if
two paths have the same suffix.